requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-filter==22.1
Markdown==3.3.7

boto3==1.24.49

python-dotenv==0.20.0

celery==5.2.7
redis==4.3.1
flower==1.0.0

psycopg2==2.9.3

aiohttp==3.8.1
requests==2.27.1


